# _refs/openapi-soundcloud-com.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "SoundCloud Public API Specification",
    "description": "Swagger json can be found [here](https://developers.soundcloud.com/docs/api/explorer/api.json)\n\n\u003Ca href=\"#\" id=\"cookie-manager\"\u003ECookie manager\u003C/a\u003E\n",
    "contact": {
      "name": "API issue tracker",
      "url": "https://github.com/soundcloud/api"
    },
    "version": "1.0.0",
    "termsOfService": "https://developers.soundcloud.com/docs/api/terms-of-use"
  },
  "servers": [
    {
      "url": "https://api.soundcloud.com"
    }
  ],
  "security": [
    {
      "AuthHeader": []
    },
    {
      "oAuth2_1": []
    }
  ],
...
  "paths": {
...
    "/users": {
      "get": {
        "summary": "Performs a user search based on a query",
        "tags": [
          "search"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/q"
          },
          {
            "$ref": "#/components/parameters/ids"
          },
          {
            "$ref": "#/components/parameters/urns"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/linked_partitioning"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Users"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/users/{user_urn}": {
      "get": {
        "summary": "Returns a user.",
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/user_urn"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/User"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/users/{user_urn}/followers": {
      "get": {
        "summary": "Returns a list of user’s followers.",
        "description": "Returns a list of users that follows (user_urn).",
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/user_urn"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Users"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users/{user_urn}/followings": {
      "get": {
        "summary": "Returns a list of user’s followings.",
        "description": "Returns list of users that (user_urn) follows.",
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/user_urn"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Users"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users/{user_urn}/followings/{following_urn}": {
      "get": {
        "summary": "Returns a user's following. (use /users/{user_urn} instead, to fetch the user details)",
        "description": "Returns (following_id) that is followed by (user_urn).",
        "deprecated": true,
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/user_urn"
          },
          {
            "$ref": "#/components/parameters/following_urn"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/User"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/users/{user_urn}/playlists": {
      "get": {
        "summary": "Returns a list of user's playlists.",
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/user_urn"
          },
          {
            "$ref": "#/components/parameters/access"
          },
          {
            "$ref": "#/components/parameters/show_tracks"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/linked_partitioning"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Playlists"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users/{user_urn}/tracks": {
      "get": {
        "summary": "Returns a list of user's tracks.",
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/user_urn"
          },
          {
            "$ref": "#/components/parameters/access"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/linked_partitioning"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Tracks"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users/{user_urn}/web-profiles": {
      "get": {
        "summary": "Returns list of user's links added to their profile (website, facebook, instagram).",
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/user_urn"
          },
          {
            "$ref": "#/components/parameters/limit"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/WebProfiles"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users/{user_urn}/likes/tracks": {
      "get": {
        "summary": "Returns a list of user's liked tracks.",
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/user_urn"
          },
          {
            "$ref": "#/components/parameters/access"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/linked_partitioning"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Tracks"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/users/{user_urn}/likes/playlists": {
      "get": {
        "summary": "Returns a list of user's liked playlists.",
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/user_urn"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/linked_partitioning"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Playlists"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "ids": {
        "in": "query",
        "name": "ids",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "A comma separated list of track ids to filter on",
        "example": "1,2,3"
      },
      "urns": {
        "in": "query",
        "name": "urns",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "A comma separated list of track urns to filter on",
        "example": "soundcloud:tracks:1,soundcloud:tracks:2,soundcloud:tracks:3"
      },
      "limit": {
        "in": "query",
        "name": "limit",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 50
        },
        "description": "Number of results to return in the collection.",
        "example": 2
      },
      "offset": {
        "deprecated": true,
        "in": "query",
        "name": "offset",
        "required": false,
        "schema": {
          "type": "integer",
          "default": 0
        },
        "description": "Offset of first result. Deprecated, use `linked_partitioning` instead.",
        "example": 0
      },
      "q": {
        "in": "query",
        "name": "q",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "search",
        "example": "hello"
      },
...
      "linked_partitioning": {
        "in": "query",
        "name": "linked_partitioning",
        "required": false,
        "schema": {
          "type": "boolean"
        },
        "description": "Returns paginated collection of items (recommended, returning a list without pagination is deprecated and should not be used)",
        "example": true
      },
...
      "access": {
        "in": "query",
        "name": "access",
        "schema": {
          "type": "array",
          "items": {
            "type": "string",
            "enum": [
              "playable",
              "preview",
              "blocked"
            ]
          },
          "default": "playable,preview"
        },
        "description": "Filters content by level of access the user (logged in or anonymous) has to the track. The result list will include only tracks with the specified access. Include all options if you'd like to see all possible tracks. See `Track#access` schema for more details.\n",
        "style": "form",
        "explode": false
      },
...
      "show_tracks": {
        "in": "query",
        "name": "show_tracks",
        "required": false,
        "schema": {
          "type": "boolean"
        },
        "description": "A boolean flag to request a playlist with or without tracks. Default is `true`.",
        "example": true
      },
...
      "following_urn": {
        "in": "path",
        "name": "following_urn",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "SoundCloud User id to denote a Following of a user",
        "example": "soundcloud:users:25219981"
      },
...
      "user_urn": {
        "in": "path",
        "name": "user_urn",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "SoundCloud User id",
        "example": "soundcloud:users:948745750"
      },
...
    },
    "securitySchemes": {
      "AuthHeader": {
        "in": "header",
        "type": "apiKey",
        "name": "Authorization",
        "description": "The string \"OAuth\" followed by the user’s token: OAuth ACCESS_TOKEN. Obtained in oAuth2.1 flow"
      },
      "oAuth2_1": {
        "type": "oauth2",
        "description": "This API uses OAuth2.1 with Authorization Code flow, Client Credential flow. When registering your app, for it to work with Swagger UI specify redirect url as https://developers.soundcloud.com/docs/api/explorer/oauth2-redirect.html. [More info](https://developers.soundcloud.com/docs/api/guide#authentication)",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://secure.soundcloud.com/authorize",
            "tokenUrl": "https://secure.soundcloud.com/oauth/token",
            "refreshUrl": "https://secure.soundcloud.com/oauth/token",
            "scopes": {

            }
          },
          "clientCredentials": {
            "tokenUrl": "https://secure.soundcloud.com/oauth/token",
            "scopes": {

            }
          }
        }
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "error": {
            "type": "string",
            "deprecated": true,
            "nullable": true
          },
          "errors": {
            "type": "array",
            "deprecated": true,
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "deprecated": true
          }
        }
      },
...
      "User": {
        "type": "object",
        "nullable": true,
        "description": "SoundCloud User object",
        "properties": {
          "avatar_url": {
            "description": "URL to a JPEG image",
            "type": "string"
          },
          "city": {
            "description": "city",
            "type": "string"
          },
          "country": {
            "description": "country",
            "type": "string"
          },
          "description": {
            "description": "description",
            "type": "string"
          },
          "discogs_name": {
            "description": "discogs name",
            "type": "string"
          },
          "first_name": {
            "description": "first name",
            "type": "string"
          },
          "followers_count": {
            "description": "number of followers",
            "type": "integer"
          },
          "followings_count": {
            "description": "number of followed users",
            "type": "integer"
          },
          "full_name": {
            "description": "first and last name",
            "type": "string"
          },
          "urn": {
            "description": "unique identifier",
            "type": "string"
          },
          "kind": {
            "description": "kind of resource",
            "type": "string"
          },
          "created_at": {
            "description": "profile creation datetime",
            "format": "date-time",
            "type": "string"
          },
          "last_modified": {
            "description": "last modified datetime",
            "format": "date-time",
            "type": "string"
          },
          "last_name": {
            "description": "last name",
            "type": "string"
          },
          "permalink": {
            "description": "permalink of the resource",
            "type": "string"
          },
          "permalink_url": {
            "description": "URL to the SoundCloud.com page",
            "type": "string"
          },
          "plan": {
            "description": "subscription plan of the user",
            "type": "string"
          },
          "playlist_count": {
            "description": "number of public playlists",
            "type": "integer"
          },
          "public_favorites_count": {
            "description": "number of favorited public tracks",
            "type": "integer"
          },
          "reposts_count": {
            "description": "number of reposts from user",
            "type": "integer"
          },
          "track_count": {
            "description": "number of public tracks",
            "type": "integer"
          },
          "uri": {
            "description": "API resource URL",
            "type": "string"
          },
          "username": {
            "description": "username",
            "type": "string"
          },
          "website": {
            "description": "a URL to the website",
            "type": "string"
          },
          "website_title": {
            "description": "a custom title for the website",
            "type": "string"
          },
          "subscriptions": {
            "description": "a list subscriptions associated with the user",
            "deprecated": true,
            "items": {
              "type": "array",
              "anyOf": [
                {
                  "description": "subscription",
                  "properties": {
                    "product": {
                      "description": "product",
                      "properties": {
                        "id": {
                          "description": "subscription id instance.",
                          "type": "string"
                        },
                        "name": {
                          "description": "subscription name instance.",
                          "type": "string"
                        }
                      },
                      "type": "object"
                    },
                    "recurring": {
                      "description": "if the subscription is recurring or not",
                      "type": "boolean"
                    }
                  },
                  "type": "object"
                }
              ]
            }
          }
        }
      },
...
      "Users": {
        "type": "object",
        "properties": {
          "collection": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "next_href": {
            "type": "string"
          }
        }
      },
...
      "Tracks": {
        "type": "object",
        "properties": {
          "collection": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Track"
            }
          },
          "next_href": {
            "type": "string"
          }
        }
      },
      "TracksList": {
        "deprecated": true,
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Track"
        }
      },
...
      "Playlists": {
        "type": "object",
        "properties": {
          "collection": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Playlist"
            }
          },
          "next_href": {
            "type": "string"
          }
        }
      },
      "PlaylistsArray": {
        "deprecated": true,
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Playlist"
        }
      },
...
      "WebProfiles": {
        "type": "array",
        "description": "User's links added to their profile",
        "items": {
          "type": "object",
          "properties": {
            "created_at": {
              "description": "Timestamp of when the link was added to the profile.",
              "type": "string"
            },
            "urn": {
              "description": "Id",
              "type": "string"
            },
            "kind": {
              "description": "Kind",
              "type": "string"
            },
            "service": {
              "description": "Service or platform",
              "type": "string"
            },
            "title": {
              "description": "Link's title",
              "type": "string"
            },
            "url": {
              "description": "URL of the external link",
              "type": "string"
            },
            "username": {
              "description": "Username extracted from the external link",
              "type": "string"
            }
          }
        }
      },
...
    },
    "responses": {
...
      "BadRequest": {
        "description": "Bad Request",
        "content": {
          "application/json; charset=utf-8": {
            "examples": {
              "default": {
                "$ref": "#/components/examples/BadRequest"
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json; charset=utf-8": {
            "examples": {
              "default": {
                "$ref": "#/components/examples/Unauthorized"
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "examples": {
              "default": {
                "$ref": "#/components/examples/NotFound"
              }
            },
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "User": {
        "description": "Success",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/User"
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/User"
              }
            }
          }
        }
      },
...
      "Users": {
        "description": "Success",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/Users"
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/Users"
              }
            }
          }
        }
      },
...
      "Tracks": {
        "description": "Success",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Tracks"
                },
                {
                  "$ref": "#/components/schemas/TracksList"
                }
              ]
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/Tracks"
              },
              "tracksList": {
                "$ref": "#/components/examples/TracksList"
              }
            }
          }
        }
      },
...
      "Playlists": {
        "description": "Success",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Playlists"
                },
                {
                  "$ref": "#/components/schemas/PlaylistsArray"
                }
              ]
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/Playlists"
              },
              "playlistsArray": {
                "$ref": "#/components/examples/PlaylistsArray"
              }
            }
          }
        }
      },
...
      "WebProfiles": {
        "description": "Success",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/WebProfiles"
            },
            "examples": {
              "default": {
                "$ref": "#/components/examples/WebProfiles"
              }
            }
          }
        }
      },
...
    },
    "examples": {
...
      "Playlists": {
        "value": {
          "collection": {
            "$ref": "#/components/examples/Playlist/value"
          },
          "next_href": "https://api.soundcloud.com/collection?page_size=10&cursor=1234567"
        }
      },
      "PlaylistsArray": {
        "description": "Deprecated, use Playlists instead.",
        "value": {
          "$ref": "#/components/examples/Playlist/value"
        }
      },
...
      "Tracks": {
        "value": {
          "collection": {
            "$ref": "#/components/examples/Track/value"
          },
          "next_href": "https://api.soundcloud.com/collection?page_size=10&cursor=1234567"
        }
      },
      "TracksList": {
        "description": "Deprecated, use Tracks instead.",
        "value": {
          "$ref": "#/components/examples/Track/value"
        }
      },
      "User": {
        "value": {
          "avatar_url": "https://i1.sndcdn.com/avatars.jpg",
          "city": "City",
          "country": "Country",
          "created_at": "2018/01/01 12:08:25 +0000",
          "description": null,
          "discogs_name": null,
          "first_name": "First_name",
          "followers_count": 0,
          "followings_count": 0,
          "full_name": "Full Name",
          "urn": "soundcloud:users:12345",
          "kind": "user",
          "last_modified": "2020/01/03 12:08:25 +0000",
          "last_name": "Last_name",
          "myspace_name": null,
          "permalink": "permalink",
          "permalink_url": "https://soundcloud.com/permalink",
          "plan": "Free",
          "playlist_count": 3,
          "public_favorites_count": 20,
          "reposts_count": 0,
          "subscriptions": [
            {
              "product": {
                "id": "some-id",
                "name": "some.name"
              }
            }
          ],
          "track_count": 0,
          "upload_seconds_left": 10800,
          "uri": "https://api.soundcloud.com/users/1234",
          "username": "some.user",
          "website": null,
          "website_title": null
        }
      },
      "Users": {
        "value": {
          "collection": {
            "$ref": "#/components/examples/User/value"
          },
          "next_href": "https://api.soundcloud.com/collection?page_size=10&cursor=1234567"
        }
      },
      "WebProfiles": {
        "value": [
          {
            "created_at": "1995/10/20 12:10:39 +0000",
            "id": 1234,
            "urn": "soundcloud:users:1234",
            "kind": "web-profile",
            "service": "facebook",
            "title": "my facebook",
            "url": "http://facebook.com",
            "username": "username"
          }
        ]
      },
...
      "BadRequest": {
        "value": {
          "code": 400,
          "message": "Detailed message of errors, when available",
          "link": "https://developers.soundcloud.com/docs/api/explorer/open-api"
        }
      },
...
      "Unauthorized": {
        "value": {
          "code": 401,
          "message": "Detailed message of errors, when available",
          "link": "https://developers.soundcloud.com/docs/api/explorer/open-api"
        }
      },
...
      "NotFound": {
        "value": {
          "code": 404,
          "message": "Detailed message of errors, when available",
          "link": "https://developers.soundcloud.com/docs/api/explorer/open-api"
        }
      },
...
    }
  }
}